move blizzard config to separate profile
