+++ compose.yaml
@@ -48,7 +48,7 @@ services:
       - SPRING_DATASOURCE_USERNAME=sc2pulse
       - COM_NEPHEST_BATTLENET_SC2_DATASOURCE_NATIVE_CLICKHOUSE_ENDPOINT=http://clickhouse:8123
       - COM_NEPHEST_BATTLENET_SC2_DATASOURCE_NATIVE_CLICKHOUSE_DATABASE=sc2pulse
-      - SPRING_PROFILES_ACTIVE=default,prod,discord,twitch
+      - SPRING_PROFILES_ACTIVE=default,prod,blizzard,discord,twitch
       - COM_NEPHEST_BATTLENET_SC2_CRON_ENABLED=true
       - COM_NEPHEST_BATTLENET_SC2_API_BLIZZARD_LADDER_ID_PREVIOUS_US=333365
       - COM_NEPHEST_BATTLENET_SC2_API_BLIZZARD_LADDER_ID_PREVIOUS_EU=268355
