+++ src/main/resources/application.properties
@@ -16,76 +16,3 @@ com.nephest.battlenet.sc2.ladder.regions=US,EU,KR
 com.nephest.battlenet.sc2.player.complaint.policy.link=https://github.com/nephest/sc2-pulse/wiki/Player-complaint-policy
 springdoc.swagger-ui.path=/doc/api
 springdoc.swagger-ui.showCommonExtensions=true
-
-spring.security.oauth2.client.provider.battlenet-global=battlenet-global
-spring.security.oauth2.client.provider.battlenet-global.issuer-uri=https://oauth.battle.net
-#spring.security.oauth2.client.provider.battlenet-cn=battlenet-cn
-#spring.security.oauth2.client.provider.battlenet-cn.issuer-uri=https://oauth.battlenet.com.cn
-
-spring.security.oauth2.client.registration.sc2-sys-us=sc2-sys-us
-spring.security.oauth2.client.registration.sc2-sys-us.provider=battlenet-global
-spring.security.oauth2.client.registration.sc2-sys-us.client-authentication-method=client_secret_basic
-spring.security.oauth2.client.registration.sc2-sys-us.authorization-grant-type=client_credentials
-spring.security.oauth2.client.registration.sc2-sys-us.client-id=${spring.security.oauth2.client.blizzard.client-id}
-spring.security.oauth2.client.registration.sc2-sys-us.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-spring.security.oauth2.client.registration.sc2-sys-eu=sc2-sys-eu
-spring.security.oauth2.client.registration.sc2-sys-eu.provider=battlenet-global
-spring.security.oauth2.client.registration.sc2-sys-eu.client-authentication-method=client_secret_basic
-spring.security.oauth2.client.registration.sc2-sys-eu.authorization-grant-type=client_credentials
-spring.security.oauth2.client.registration.sc2-sys-eu.client-id=${spring.security.oauth2.client.blizzard.client-id}
-spring.security.oauth2.client.registration.sc2-sys-eu.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-spring.security.oauth2.client.registration.sc2-sys-kr=sc2-sys-kr
-spring.security.oauth2.client.registration.sc2-sys-kr.provider=battlenet-global
-spring.security.oauth2.client.registration.sc2-sys-kr.client-authentication-method=client_secret_basic
-spring.security.oauth2.client.registration.sc2-sys-kr.authorization-grant-type=client_credentials
-spring.security.oauth2.client.registration.sc2-sys-kr.client-id=${spring.security.oauth2.client.blizzard.client-id}
-spring.security.oauth2.client.registration.sc2-sys-kr.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-#spring.security.oauth2.client.registration.sc2-sys-cn=sc2-sys-cn
-#spring.security.oauth2.client.registration.sc2-sys-cn.provider=battlenet-cn
-#spring.security.oauth2.client.registration.sc2-sys-cn.client-authentication-method=client_secret_basic
-#spring.security.oauth2.client.registration.sc2-sys-cn.authorization-grant-type=client_credentials
-#spring.security.oauth2.client.registration.sc2-sys-cn.client-id=${spring.security.oauth2.client.blizzard.client-id}
-#spring.security.oauth2.client.registration.sc2-sys-cn.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-spring.security.oauth2.client.registration.sc2-lg-eu=sc2-lg-eu
-spring.security.oauth2.client.registration.sc2-lg-eu.provider=battlenet-global
-spring.security.oauth2.client.registration.sc2-lg-eu.client-authentication-method=client_secret_basic
-spring.security.oauth2.client.registration.sc2-lg-eu.authorization-grant-type=authorization_code
-spring.security.oauth2.client.registration.sc2-lg-eu.scope=openid
-spring.security.oauth2.client.registration.sc2-lg-eu.redirect-uri=${com.nephest.battlenet.sc2.url.public:localhost}login/oauth2/code/battlenet-eu
-spring.security.oauth2.client.registration.sc2-lg-eu.client-name=BattleNet EU
-spring.security.oauth2.client.registration.sc2-lg-eu.client-id=${spring.security.oauth2.client.blizzard.client-id}
-spring.security.oauth2.client.registration.sc2-lg-eu.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-spring.security.oauth2.client.registration.sc2-lg-us=sc2-lg-us
-spring.security.oauth2.client.registration.sc2-lg-us.provider=battlenet-global
-spring.security.oauth2.client.registration.sc2-lg-us.client-authentication-method=client_secret_basic
-spring.security.oauth2.client.registration.sc2-lg-us.authorization-grant-type=authorization_code
-spring.security.oauth2.client.registration.sc2-lg-us.scope=openid
-spring.security.oauth2.client.registration.sc2-lg-us.redirect-uri=${com.nephest.battlenet.sc2.url.public:localhost}login/oauth2/code/battlenet-us
-spring.security.oauth2.client.registration.sc2-lg-us.client-name=BattleNet US
-spring.security.oauth2.client.registration.sc2-lg-us.client-id=${spring.security.oauth2.client.blizzard.client-id}
-spring.security.oauth2.client.registration.sc2-lg-us.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-spring.security.oauth2.client.registration.sc2-lg-kr=sc2-lg-kr
-spring.security.oauth2.client.registration.sc2-lg-kr.provider=battlenet-global
-spring.security.oauth2.client.registration.sc2-lg-kr.client-authentication-method=client_secret_basic
-spring.security.oauth2.client.registration.sc2-lg-kr.authorization-grant-type=authorization_code
-spring.security.oauth2.client.registration.sc2-lg-kr.scope=openid
-spring.security.oauth2.client.registration.sc2-lg-kr.redirect-uri=${com.nephest.battlenet.sc2.url.public:localhost}login/oauth2/code/battlenet-kr
-spring.security.oauth2.client.registration.sc2-lg-kr.client-name=BattleNet KR
-spring.security.oauth2.client.registration.sc2-lg-kr.client-id=${spring.security.oauth2.client.blizzard.client-id}
-spring.security.oauth2.client.registration.sc2-lg-kr.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
-
-#spring.security.oauth2.client.registration.sc2-lg-cn=sc2-lg-cn
-#spring.security.oauth2.client.registration.sc2-lg-cn.provider=battlenet-cn
-#spring.security.oauth2.client.registration.sc2-lg-cn.client-authentication-method=client_secret_basic
-#spring.security.oauth2.client.registration.sc2-lg-cn.authorization-grant-type=authorization_code
-#spring.security.oauth2.client.registration.sc2-lg-cn.scope=openid
-#spring.security.oauth2.client.registration.sc2-lg-cn.redirect-uri=${com.nephest.battlenet.sc2.url.public:localhost}login/oauth2/code/battlenet-cn
-#spring.security.oauth2.client.registration.sc2-lg-cn.client-name=BattleNet CN
-#spring.security.oauth2.client.registration.sc2-lg-cn.client-id=${spring.security.oauth2.client.blizzard.client-id}
-#spring.security.oauth2.client.registration.sc2-lg-cn.client-secret=${spring.security.oauth2.client.blizzard.client-secret}
