+++ src/main/resources/templates/fragments/main-section.html
@@ -363,7 +363,9 @@
 </section>
 <div th:fragment="login">
     <ul class="d-flex flex-center-wrap-gap list-unstyled">
-        <th:block th:with="activeRegions=${@environment.getProperty('com.nephest.battlenet.sc2.ladder.regions').split(',')}">
+        <th:block th:if="${@environment.acceptsProfiles('blizzard')}"
+                  th:with="activeRegions=${@environment.getProperty('com.nephest.battlenet.sc2.ladder.regions').split(',')}"
+        >
             <li th:if="${#arrays.contains(activeRegions, 'US') or #arrays.contains(activeRegions, 'EU') or #arrays.contains(activeRegions, 'KR')}">
                 <a
                         class="btn btn-outline-primary"
