+++ src/test/java/com/nephest/battlenet/sc2/model/blizzard/BlizzardTest.java
@@ -1,0 +1,21 @@
+// Copyright (C) 2020-2026 Oleksandr Masniuk
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+package com.nephest.battlenet.sc2.model.blizzard;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.test.context.junit.jupiter.EnabledIf;
+
+@EnabledIf
+(
+    expression = "#{environment.acceptsProfiles('blizzard')}",
+    reason = "'blizzard' profile is disabled",
+    loadContext = true
+)
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface BlizzardTest
+{}
